feat(color-picker): 'compose可拆分函数'
